requirements.txt
Flask
gunicorn
numpy
pillow
pandas
nltk
stopwords
scikit-learn
imbalanced-learn
pyyaml
mlflow
hydra-core